refactor: List 컴포넌트 디렉토리 네이밍 수정
- 디렉토리명 List/NameList 컴포넌트명 NameList -> 디렉토리명 List/Name 컴포넌트명 NameList
