+++ frontend/src/components/List/Story/style.ts
